requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.6.1
requests==2.32.5
Werkzeug==3.1.4
python-socketio==5.16.1
python-engineio==4.13.1
chardet==7.4.0.post2
flask-cors==6.0.2
PyYAML==6.0.2
